% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.data.R
\name{check.data}
\alias{check.data}
\title{Check DATA.bib Entries}
\usage{
check.data(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to suppress messages.}
}
\value{
Logical vector indicating which entries have been processed.

A warning is generated if any entries have not been processed.
}
\description{
Check if all \verb{DATA.bib} entries have been processed.
}
\examples{
\dontrun{
check.data()
check.data(quiet=TRUE)
}

}
\seealso{
\code{\link{check.software}} checks \verb{SOFTWARE.bib} versions.

\code{\link{taf.boot}} runs the boot procedure and sets up data files
according to \verb{DATA.bib} specifications.

\code{\link{TAF-package}} gives an overview of the package.
}
