% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.software.R
\name{check.software}
\alias{check.software}
\title{Check versions with those in SOFTWARE.bib}
\usage{
check.software()
}
\value{
TRUE or FALSE if available packages match those required or not.
}
\description{
Compare installed and required package versions
}
\details{
A SOFTWARE.bib file can be use to record what package versions a certain
analysis expects. Those available in the running session can be compared
with those listed in the file. A warning is printed if the former are
earlier that the later.
}
\seealso{
[draft.software()]
}
\keyword{classes}
