% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{deps}
\alias{deps}
\title{Dependencies of a Workflow}
\usage{
deps(path = ".", base = FALSE, installed = TRUE, available = TRUE,
  list = FALSE)
}
\arguments{
\item{path}{a directory or file containing R code.}

\item{base}{whether to include base packages in the output.}

\item{installed}{whether to include installed packages in the output.}

\item{available}{whether to include available packages in the output.}

\item{list}{whether to return packages in list format, split by script.}
}
\value{
Names of packages as a vector, or in list format if \code{list=TRUE}. If no
dependencies are found, the return value is \code{NULL}.
}
\description{
Scan R scripts for packages that are required.
}
\details{
The files scanned are those with the file extensions \code{.R}, \code{.Rmd},
\code{Rnw}, and \code{qmd}.
}
\note{
Package names are matched based on four patterns:\preformatted{
library(*)
require(*)
*::object
*:::object}

The scanning algorithm may return false-positive dependencies if these
patterns occur inside if-clauses, strings, comments, etc.
}
\examples{
\dontrun{
dir <- system.file(package="MASS", "scripts")
script <- system.file(package="MASS", "scripts/ch08.R")

deps(script)                   # dependencies
deps(script, base=TRUE)        # including base packages
deps(script, installed=FALSE)  # not (yet) installed

deps(dir)
deps(dir, list=TRUE)

deps(dir, available=FALSE)  # dependencies that might be unavailable
}

}
\seealso{
\code{\link{installed.packages}}, \code{\link{available.packages}}.

\code{\link{pdeps}} shows the dependencies of a package.

\code{\link{TAF-package}} gives an overview of the package.
}
