% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.readme.R
\name{draft.readme}
\alias{draft.readme}
\title{Draft Readme}
\usage{
draft.readme(path = ".", title = NULL, force = FALSE)
}
\arguments{
\item{path}{directory of a TAF analysis.}

\item{title}{title to display at the top of the `README.md` page.}

\item{force}{whether to overwrite an existing \code{README.md} file.}
}
\value{
No return value, called for side effects.
}
\description{
Draft an introductory \code{README.md} that describes how to run a TAF
analysis.
}
\details{
The default value of \code{title = NULL} uses the Git repository name as a
placeholder title.
}
\examples{
\dontrun{
draft.readme()
}

}
\seealso{
\code{\link{TAF-package}} gives an overview of the package.
}
