% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdim.R
\name{fdim}
\alias{fdim}
\title{FLR Dimensions}
\usage{
fdim(x, reduce = FALSE)
}
\arguments{
\item{x}{an FLR table in \code{FLQuant} format or in a 7-column
\code{data.frame} format.}

\item{reduce}{is whether to omit single-level dimensions.}
}
\value{
Named vector showing the dimension names and number of levels.
}
\description{
Show the dimensions of an FLR table.
}
\details{
\code{x} can also be an arbitrary data frame (not related to FLR) where the
last column is the measurement variable.
}
\examples{
\dontrun{
library(FLCore)
flq <- FLQuant(t(catage.xtab), dimnames=list(age=1:4, year=1963:1970))
df <- as.data.frame(flq)
fdim(flq)
fdim(df)
fdim(df, TRUE)
}

# Any data frame where the last column is the measurement variable
fdim(catage.long)
fdim(esoph[-5])
fdim(rev(warpbreaks))
fdim(rev(ToothGrowth))

}
\seealso{
\code{\link{dim}}, \code{\link{dimnames}}, and \code{\link{names}} are the
base functions to access dimension names and levels.

\code{\link{as.data.frame}} is a method provided by the \pkg{FLCore} package
to convert \code{FLQuant} tables to a 7-column long format.

\code{\link{flr2taf}} converts an FLR table to TAF format.

\code{\link{TAF-package}} gives an overview of the package.
}
