% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.name.R
\name{git.name}
\alias{git.name}
\title{Git Name}
\usage{
git.name(path = ".", owner = TRUE, warn = TRUE)
}
\arguments{
\item{path}{top directory of a Git repository.}

\item{owner}{whether to include the repository owner name.}

\item{warn}{whether to raise a warning include the repository owner name.}
}
\value{
String of the format \code{"[owner]/repo"}.
}
\description{
Read the Git repository name from the \code{.git} folder.
}
\examples{
\dontrun{
git.name()
git.name(owner=FALSE)
}

}
\seealso{
\code{\link{TAF-package}} gives an overview of the package.
}
