% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.taf.R
\name{make.taf}
\alias{make.taf}
\alias{makeTAF}
\title{Run TAF Script if Needed}
\usage{
make.taf(script, ...)
}
\arguments{
\item{script}{TAF script filename.}

\item{...}{passed to \code{\link{make}} and \code{\link{source.taf}}.}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether the script was run.
}
\description{
Run a TAF script if the target directory is either older than the script, or
older than the directory of the previous TAF step.
}
\note{
Any underlying scripts are automatically included if they share the same
filename prefix, followed by an underscore. For example, when determining
whether a script \verb{data.R} should be run, this function checks whether
\verb{data_foo.R} and \verb{data_bar.R} have been recently modified.
}
\examples{
\dontrun{
make.taf("model.R")
}

}
\seealso{
\code{\link{source}} runs any R script, \code{\link{source.taf}} is more
convenient for running a TAF script, and \code{\link{source.all}} runs all
TAF scripts.

\code{\link{make}}, \code{\link{make.taf}}, and \code{\link{make.all}} are
similar to the \code{source} functions, except they avoid repeating tasks
that have already been run.

\code{\link{TAF-package}} gives an overview of the package.
}
