% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Show Message}
\usage{
msg(...)
}
\arguments{
\item{\dots}{passed to \code{message}.}
}
\value{
No return value, called for side effects.
}
\description{
Show a message, as well as the current time.
}
\examples{
\donttest{
msg("script.R running...")
}

}
\seealso{
\code{\link{message}} is the base function to show messages, without the
current time.

\code{\link{source.taf}} reports progress using \code{msg}.

\code{\link{TAF-package}} gives an overview of the package.
}
