% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bib.R
\name{read.bib}
\alias{read.bib}
\title{Bibtex Parser}
\usage{
read.bib(file)
}
\arguments{
\item{file}{\file{*.bib} file to parse.}
}
\value{
List containing Bibtex entries.
}
\description{
Parser for Bibtex entries.
}
\note{
This parser is similar to the \code{\link[bibtex]{read.bib}} function in the
\pkg{bibtex} package, except:
\itemize{
\item It returns a plain list instead of class \code{bibentry}.
\item The fields \code{bibtype} and \code{key} are stored as plain list
      elements instead of attributes.
}
}
\seealso{
\code{\link{taf.sources}} calls this function to import metadata entries.

\code{\link{TAF-package}} gives an overview of the package.
}
