% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.list.R
\name{read.list}
\alias{read.list}
\title{Read List}
\usage{
read.list(file, prefix = "# ")
}
\arguments{
\item{file}{a filename.}

\item{prefix}{a string prefix before element names.}
}
\value{
A list.
}
\description{
Read a list from a text file.
}
\note{
The \code{\link{write.list}} function can be useful to produce a
human-readable text file containing a simple list, which can be imported back
into R using \code{read.list}.

Other pathways, such as \code{\link{saveRDS}}, \code{\link{dput}}, or JSON,
are better suited for large or complex lists when human readability is not a
priority.
}
\examples{
x <- list(pi=pi, month=month.name, value=stack.loss,
          country=c("Bay of Biscay", "Kattegat", "North Sea"))

write.list(x)
write.list(x, ncolumns=NULL)
write.list(x, ncolumns=3)
write.list(x, ncolumns=3, sep=",")
write.list(x, prefix="# ")

\dontrun{
write.list(x, "list.dat")
}

}
\seealso{
\code{\link{write.list}} writes a list to a text file.

\code{\link{write}} is the underlying function used to write data values to a
file.

\code{\link{TAF-package}} gives an overview of the package.
}
