% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.example.R
\name{taf.example}
\alias{taf.example}
\title{TAF Example}
\usage{
taf.example(name = "linreg", path = ".", force = FALSE,
  gitignore = FALSE)
}
\arguments{
\item{name}{of TAF example analysis.}

\item{path}{where to create example directory. The default is the current
working directory.}

\item{force}{whether to overwrite existing directory.}

\item{gitignore}{whether to write TAF entries to a \file{.gitignore} file.}
}
\value{
Full path to directory that was created.
}
\description{
Copy example analysis from TAF package.
}
\details{
When \code{gitignore = TRUE}, the following entries will be written to a
\verb{.gitignore} file:
\preformatted{
/boot/data
/boot/library
/boot/software
/data
/model
/output
/report
*.Rproj
.RData
.Rhistory
.Rproj.user
.Ruserdata
}
}
\note{
The example analysis is copied from the TAF package directory:
\preformatted{
dir(system.file("examples", package="TAF"), full=TRUE)
}
}
\examples{
\dontrun{
taf.example()
}

}
\seealso{
\code{\link{taf.skeleton}} creates an empty TAF template.

\code{\link{TAF-package}} gives an overview of the package.
}
