% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.lib.path.R
\name{taf.lib.path}
\alias{taf.lib.path}
\title{Construct Path to the TAF boot library}
\usage{
taf.lib.path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors. Long vectors are not supported.}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\value{
character
}
\description{
Construct the path to the TAF boot library folder
from components in a platform-independent way.  This function
checks to see if R is running in the boot folder - i.e.
`taf.boot()` is running, and adjusts the path accordingly.
}
\details{
This function, simplifies the construction of file paths to
inintial data files gathered during the TAF booting step.
In addition, this function is useful when developing scripts used
in the boot procedure, as these scripts are run with the
working directory set to the boot folder, and hence make it
to develop and debug.
}
\seealso{
\link{file.path}
}
