% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.libraries.R
\name{taf.libraries}
\alias{taf.libraries}
\title{TAF Libraries}
\usage{
taf.libraries(messages = FALSE, warnings = FALSE)
}
\arguments{
\item{messages}{whether to show messages when package loads.}

\item{warnings}{whether to show warnings when package loads.}
}
\value{
\code{TRUE} (invisibly) if all packages loaded.
}
\description{
Load and attach all packages from local TAF library.
}
\note{
Packages in the TAF library are loaded in the order in which they are
listed in \verb{SOFTWARE.bib}. Internal dependencies can in this way be
respected.
}
\examples{
\dontrun{
# Load all packages in TAF library
taf.libraries()
}

}
\seealso{
\code{\link{taf.library}} is the TAF function called for each found package.

\code{\link{TAF-package}} gives an overview of the package.
}
