% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.list.R
\name{write.list}
\alias{write.list}
\title{Write List}
\usage{
write.list(x, file = "", ncolumns = NULL, append = FALSE, sep = " ",
  section = "")
}
\arguments{
\item{x}{a list containing atomic elements.}

\item{file}{a filename, or special value \code{""}.}

\item{ncolumns}{the number of columns to write the data in.}

\item{append}{whether to append to an existing file.}

\item{sep}{a string used to separate columns.}

\item{section}{a string to use before section labels.}
}
\value{
No return value, called for side effects.
}
\description{
Write a list to a text file.
}
\details{
The default value \code{file = ""} prints the list in the console, allowing
the user to preview alternative formats before writing the list to a file.

The default value \code{ncolumns = NULL} uses 1 column for strings and 5
columns for other data types, in the same way as \code{\link{write}} does.
}
\examples{
x <- list(a=pi, b=month.name, c=stack.loss)

write.list(x)
write.list(x, ncolumns=c(1,2,3))
write.list(x, sep=",")
write.list(x, section="# ")

\dontrun{
write.list(x, "list.dat")
}

}
\seealso{
\code{\link{write}} is the underlying function used to write data values to a
file.

\code{\link{TAF-package}} gives an overview of the package.
}
