% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.tree.R
\name{dir.tree}
\alias{dir.tree}
\title{Directory Tree}
\usage{
dir.tree(path = ".", style = "unix")
}
\arguments{
\item{path}{the directory to show.}
}
\description{
Show directory structure and file contents in a tree format.
}
\examples{
\dontrun{
path <- system.file("examples", package="TAF")
dir.tree(path)
}

}
\seealso{
\link{dir} is the underlying base function that returns directories and files
as strings.

\code{\link{TAF-package}} gives an overview of the package.
}
