% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAF-internal.R, R/boot.dir.R, R/boot.exists.R,
%   R/download.github.R, R/ds.file.R, R/ds.package.R, R/model.dir.R,
%   R/model.script.R, R/parse.repo.R, R/process.bibfile.R, R/process.entry.R,
%   R/taf.install.R, R/taf.skeleton.R
\name{TAF-internal}
\alias{TAF-internal}
\alias{boot.dir}
\alias{boot.dir.inside}
\alias{boot.exists}
\alias{extract.subdir}
\alias{stamp.description}
\alias{ds.file}
\alias{ds.package}
\alias{model.dir}
\alias{model.script}
\alias{parse.repo}
\alias{process.bibfile}
\alias{process.entry}
\alias{already.in.taf.library}
\alias{taf.pkg}
\title{Various TAF Helper Functions}
\usage{
boot.dir()

boot.dir.inside(inside = ".")

boot.exists()

extract.subdir(targz, subtargz, subdir)

stamp.description(targz, spec, sha.full)

ds.file(package, author, year, title, version, source)

ds.package(package, author, year, title, version, source)

model.dir()

model.script()

parse.repo(repo)

process.bibfile(type, clean = TRUE, quiet = FALSE, ...)

process.entry(bib, clean = FALSE, comma.split = TRUE, force = FALSE,
  quiet = FALSE)

already.in.taf.library(targz, lib)

taf.pkg()
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{TAF-package}} gives an overview of the package.
}
\keyword{internal}
