% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.skeleton.R
\name{taf.skeleton}
\alias{taf.skeleton}
\title{TAF Skeleton}
\usage{
taf.skeleton(path = ".", force = FALSE, pkgs = "TAF")
}
\arguments{
\item{path}{where to create initial directories and R scripts. The default is
the current working directory.}

\item{force}{whether to overwrite existing scripts.}

\item{pkgs}{packages to load at the start of each script. The default is the
TAF package, i.e. \code{library(TAF)}.}
}
\value{
Full path to analysis directory.
}
\description{
Create initial directories and R scripts for a new TAF analysis.
}
\examples{
\dontrun{
taf.skeleton()
}

}
\seealso{
\code{\link{package.skeleton}} creates an empty template for a new R package.

\code{\link{TAF-package}} gives an overview of the package.
}
