% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.list.R
\name{write.list}
\alias{write.list}
\title{Write List}
\usage{
write.list(x, file = "", ncolumns = 1, sep = " ", prefix = "# ",
  as.is = FALSE)
}
\arguments{
\item{x}{a list containing atomic elements.}

\item{file}{a filename, or special value \code{""}.}

\item{ncolumns}{the number of columns to write the data in.}

\item{sep}{a string used to separate columns.}

\item{prefix}{a string prefix to use before element names.}

\item{as.is}{whether to simply capture the output of printing \code{x} to the
console in standard R format and write that to the \code{file}.}
}
\value{
No return value, called for side effects.
}
\description{
Write a list to a text file.
}
\details{
The default value \code{file = ""} prints the list in the console, allowing
the user to preview alternative formats before writing the list to a file.

The default value \code{ncolumns = 1} writes one data value per line. This is
the only format supported by \code{\link{read.list}} and therefore
recommended if importing the list back into R is relevant. Other formats
using \code{ncolumns} and \code{as.is} can improve human readability if
importing back into R is not relevant.

The special value \code{ncolumns = NULL} uses 1 column for strings and 5
columns for other data types, in the same way as \code{\link{write}} does.

The \code{as.is = TRUE} format is affected by R session options such as
\code{width} and \code{digits}, but is not affected by the function arguments
\code{ncolumns}, \code{sep}, and \code{prefix}.
}
\note{
The \code{write.list} function can be useful to produce a human-readable text
file containing a simple list, which can be imported back into R using
\code{\link{read.list}}.

Other pathways, such as \code{\link{saveRDS}}, \code{\link{dput}}, or JSON,
are better suited for large or complex lists when human readability is not a
priority.
}
\examples{
x <- list(pi=pi, month=month.name, value=stack.loss,
          area=c("Bay of Biscay", "Kattegat", "North Sea"))

write.list(x)
write.list(x, ncolumns=NULL)
write.list(x, ncolumns=3)
write.list(x, ncolumns=3, sep=",")
write.list(x, prefix="$")

\dontrun{
write.list(x, "list.dat")
}

}
\seealso{
\code{\link{read.list}} reads a list from a text file.

\code{\link{write}} is the underlying function used to write data values to a
file.

\code{\link{TAF-package}} gives an overview of the package.
}
