% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.list.R
\name{write.list}
\alias{write.list}
\title{Write List}
\usage{
write.list(x, file = "", ncolumns = 1, sep = " ", prefix = "")
}
\arguments{
\item{x}{a list containing atomic elements.}

\item{file}{a filename, or special value \code{""}.}

\item{ncolumns}{the number of columns to write the data in.}

\item{sep}{a string used to separate columns.}

\item{prefix}{a string prefix to use before element names.}
}
\value{
No return value, called for side effects.
}
\description{
Write a list to a text file.
}
\details{
The default value \code{file = ""} prints the list in the console, allowing
the user to preview alternative formats before writing the list to a file.

The default value \code{ncolumns = 1} writes one data value per line. This is
the only format supported by \code{\link{read.list}} and therefore
recommended if importing the list back into R is relevant. Other formats
using \code{ncolumns} and \code{sep} can improve human readability if
importing back into R is not relevant.

The special value \code{ncolumns = NULL} uses 1 column for strings and 5
columns for other data types, in the same way as \code{\link{write}} does.
}
\examples{
x <- list(pi=pi, month=month.name, value=stack.loss,
          area=c("Bay of Biscay", "Kattegat", "North Sea"))

write.list(x)
write.list(x, ncolumns=NULL)
write.list(x, ncolumns=3)
write.list(x, ncolumns=3, sep=",")
write.list(x, prefix="# ")

\dontrun{
write.list(x, "list.dat")
}

}
\seealso{
\code{\link{write}} is the underlying function used to write data values to a
file.

\code{\link{TAF-package}} gives an overview of the package.
}
